pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python_pubsub_devtools"
version = "0.2.0"
description = "Generic development and debugging tools for PubSub-based event-driven architectures - Event Flow Visualization, Event Recording, and Scenario Testing with Chaos Engineering"
readme = "README.md"
authors = [
    { name = "venantvr", email = "venantvr@gmail.com" }
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "flask>=2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pandas>=2.0.0",
    "pydot>=1.4.0",
    "matplotlib>=3.5.0",
    "networkx>=2.8.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "setuptools>=65.0",
    "wheel",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "black>=23.0",
]

[project.scripts]
pubsub-devtools = "python_pubsub_devtools.cli.main:cli"

[project.urls]
Homepage = "https://github.com/venantvr-trading/Python.PubSub.DevTools"
Documentation = "https://github.com/venantvr-trading/Python.PubSub.DevTools/wiki"
Repository = "https://github.com/venantvr-trading/Python.PubSub.DevTools"
Issues = "https://github.com/venantvr-trading/Python.PubSub.DevTools/issues"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["python_pubsub_devtools*"]
exclude = []
namespaces = false

[tool.setuptools.package-data]
python_pubsub_devtools = [
    "web/templates/*.html",
    "web/static/css/*.css",
    "web/static/js/*.js",
    "web/static/*.svg",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.black]
line-length = 100
target-version = ['py310']
